LibCards: Draw card values in the middle of the card
This is just a test and should not be merged anywhere. First off,
Painter::draw_upsidedown_bitmap should no be it's own function. It should
be consolidated into some other function.

The middle symbols should be larger than the corner ones. Currently the
symbols can't be centered b/c their width is an odd number and the card
width is an even number (80).

Also, aces, jacks, queens, and kings should have custom bitmaps.